Move boundary condition to 4C
