+++ tests/test_four_c.py
@@ -31,13 +31,13 @@ input files."""
 
 import numpy as np
 
-from meshpy.core.boundary_condition import BoundaryCondition
 from meshpy.core.conf import mpy
 from meshpy.core.function import Function
 from meshpy.core.geometry_set import GeometrySet
 from meshpy.core.material import MaterialReissner
 from meshpy.core.rotation import Rotation
 from meshpy.four_c.beam_potential import BeamPotential
+from meshpy.four_c.boundary_condition import BoundaryCondition
 from meshpy.four_c.dbc_monitor import linear_time_transformation
 from meshpy.four_c.element_beam import Beam3rHerm2Line3
 from meshpy.four_c.input_file import InputFile
